// pwa/src/generated/models/groupInfoResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */
import type { GroupKind } from './groupKind';
import type { GroupRole } from './groupRole';
import type { GroupVisibility } from './groupVisibility';
import type { MemberSummary } from './memberSummary';
import type { SnowflakeID } from './snowflakeID';

export interface GroupInfoResponse {
  avatar?: string;
  avatarImageId?: SnowflakeID;
  createdAt: string;
  description?: string;
  id: SnowflakeID;
  kind: GroupKind;
  mutedUntil?: string;
  myRole?: GroupRole;
  name: string;
  /** The other participant, for DM chats; `None` for regular groups. */
  peer?: MemberSummary;
  visibility: GroupVisibility;
}